refactor: Supabaseベストプラクティスに従ってテストファイル構造を再編成
- 共有ファイルを_shared/ディレクトリに移動（Supabase推奨構造）
- テストファイルを__tests__/ディレクトリに分離
- インポートパスを新しい構造に合わせて更新

この構造により以下が改善：
- Supabase標準の_shared/ディレクトリ利用
- テストファイルの明確な分離
- デプロイ時のテストファイル除外が容易
- 複数Edge Functions間での共有コード管理
