+++ supabase/functions/_shared/test-helpers/gemini-mock.ts
