+++ supabase/functions/process-pdf-single/__tests__/promptRegistry.test.ts
